refactor: create import path alias
