+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './ChallengeBox';
+export * from './CompletedChallenges';
+export * from './CountDown';
+export * from './ExperienceBar';
+export * from './LevelUpModal';
+export * from './Profile';
