+++ src/hooks/useChallengeContext.hook.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ChallengeContext, type ChallengeContextInterface } from '../contexts';
+import { ChallengeContext, type ChallengeContextInterface } from '~/contexts';
 
 export default function useChallengeContext(): ChallengeContextInterface {
   return useContext(ChallengeContext);
