+++ src/hooks/useCountDownContext.hook.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { CountDownContext, type CountDownContextInterface } from '../contexts';
+import { CountDownContext, type CountDownContextInterface } from '~/contexts';
 
 export default function useCountDownContext(): CountDownContextInterface {
   return useContext(CountDownContext);
