+++ tsconfig.json
@@ -14,6 +14,12 @@
     "module": "esnext",
     "moduleResolution": "node",
     "noEmit": true,
+    "baseUrl": "src/",
+    "paths": {
+      "~/*": [
+        "*"
+      ]
+    },
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "strict": false,
